pyproject.toml
[tool.poetry]
name = "django-rest-framework"
version = "0.1.0"
description = ""
authors = ["Xaron11 <kaczmarzykdawid4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
PyJWT = "^2.1.0"
python-decouple = "^3.4"
django-debug-toolbar = "^3.2.1"
django-seed = "^0.2.2"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
